tests/wpt/tests/svg/fonts/zero-font-size-adjust-no-glyphs-font.html

<!DOCTYPE html>
<title>Verify effect of font-size-adjust: 0 on a SVG Font without glyphs</title>
<link rel="help" href="https://drafts.csswg.org/css-fonts-5/#font-size-adjust-prop">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  pre { font-size-adjust: 0; font-family: CustomMonospace }
</style>
<svg>
  <font>
    <font-face font-family="CustomMonospace"/>
  </font>
</svg>
<pre id="pre">Text
</pre>
<script>
  test((t) => {
    assert_equals(pre.getBoundingClientRect().height, 0);
  }, "Verify effect of font-size-adjust: 0 on no glyphs SVG Font");
</script>
